feat: add support for setting URL parameters on client requests
Several NS1 API endpoints allow setting URL parameters to alter the
behavior of the endpoint. For instance, a `limit` parameter can be set
on the `/account/activity` endpoint to control how many events are
returned in the response. This would be useful in #233.

This commit adds support for URL parameters by adding a new struct
`Param` to the `rest` package, and updating `rest.client.Do()` and
`rest.client.DoWithPagination()` to accept a variable amount of params.
In order for higher level services based off of the common service to
set url params, they will need to pass them through when invoking the
client's `Do()` method, like so:

```golang
resp, err := s.client.Do(req, &al, params...)
```